fix(catalog): a fractional communication ratio must not be truncated on the wire
comm_ratio and ondevice_ratio were typed Integer in ModelRecipeDto. That held
while every measurement happened to be a large whole number -- the chest X-ray
campaign's 3,321x -- and broke the moment PNEUMONIA_CNN was measured on the
product path at 1.004x, because that recipe's classifier is 99.6% of its
parameters so freezing saves almost nothing.

Jackson coerces a JSON float into an Integer by TRUNCATING, so the honest 1.004
would have reached the picker as 1: a measured value silently rounded toward a
rounder-looking claim, on exactly the number that makes the frozen arm look
pointless for that recipe. Both fields are now Double.

Tests cover the three cases that matter: a fractional ratio survives intact, a
large whole ratio still parses, and null stays null rather than becoming 0 --
"not measured" and "no saving measured" are different statements.

Found by checking the cross-layer shape after the picker moved to per-recipe
measurements, not by a failing test; the existing suite passed throughout because
no fixture had ever carried a fraction.

backend full suite + jacoco gate green.

+++ backend/fl-platform-api/src/main/java/com/federated/fl_platform_api/dto/ModelRecipeDto.java
@@ -55,8 +55,16 @@ public record ModelRecipeDto(
     @JsonIgnoreProperties(ignoreUnknown = true)
     public record ArmTradeoff(
             @JsonProperty("headline") String headline,
-            @JsonProperty("commRatio") @JsonAlias("comm_ratio") Integer commRatio,
-            @JsonProperty("ondeviceRatio") @JsonAlias("ondevice_ratio") Integer ondeviceRatio,
+            /**
+             * Communication saving from freezing, as a ratio. {@code Double}, not {@code Integer}:
+             * these were all large whole numbers (3,321x) until PNEUMONIA_CNN was measured on the
+             * product path at <b>1.004x</b> — that recipe's classifier is 99.6% of its parameters,
+             * so freezing saves almost nothing. Jackson coerces a JSON float into an Integer by
+             * TRUNCATING, which would have rounded an honest 1.004 to 1 on its way to the picker.
+             * {@code null} means not measured, and must never become 0.
+             */
+            @JsonProperty("commRatio") @JsonAlias("comm_ratio") Double commRatio,
+            @JsonProperty("ondeviceRatio") @JsonAlias("ondevice_ratio") Double ondeviceRatio,
             @JsonProperty("measuredOn") @JsonAlias("measured_on") java.util.Map<String, String> measuredOn,
             @JsonProperty("arms") java.util.Map<String, java.util.Map<String, Object>> arms,
             @JsonProperty("caveats") List<String> caveats
